pyproject.toml
[project]
name = "embaiml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyterlab>=4.5.3",
    "pyserial>=3.5",
]
